Add tests of for comprehension bindings
This also splits up and organizes the tests by zero/single/multiple use.
