+++ input/src/main/scala/fix/Box.scala
@@ -1,0 +1,8 @@
+/*
+rule = LinearTypes
+*/
+package fix
+
+import com.earldouglas.linearscala.Linear
+
+case class Box(value: Int) extends Linear
