+++ input/src/main/scala/fix/SingleUse.scala
@@ -1,0 +1,9 @@
+/*
+rule = LinearTypes
+*/
+package fix
+
+trait FieldUsedOnce {
+  val box: Box = Box(42)
+  println(box)
+}
